~0U
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@1607 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/EM.cpp
@@ -2186,7 +2186,7 @@ DECLINLINE(int) emR3RawHandleRC(PVM pVM, PCPUMCTX pCtx, int rc)
                 rc = TRPMQueryTrapAll(pVM, &u8Interrupt, NULL, &uErrorCode, NULL);
                 AssertRC(rc);
 
-                if (uErrorCode != (uint32_t)~0)
+                if (uErrorCode != ~0U)
                     enmError = TRPM_TRAP_HAS_ERRORCODE;
 
                 /* If the guest gate is marked unpatched, then we will check again if we can patch it. */
